personal-notes-bot/requirements.txt
langchain>=0.0.300
chromadb>=0.3.29
sentence-transformers>=2.2.2
transformers>=4.35.0
accelerate>=0.21.0
pdfplumber>=0.8.0
python-docx>=0.8.11
streamlit>=1.25.0
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pydantic>=1.10.0
python-dotenv>=1.0.0
requests>=2.28.0
